+++ R/az_dir_role.R
@@ -14,6 +14,7 @@
 #' - `update(...)`: Update the item's properties in Microsoft Graph.
 #' - `do_operation(...)`: Carry out an arbitrary operation on the item.
 #' - `sync_fields()`: Synchronise the R object with the item metadata in Microsoft Graph.
+#' - `list_members(n=Inf)`: Return a list of all members of this group. `n` is the number of results to return; set this to NULL to return the `ms_graph_pager` iterator object for the result set.
 #'
 #' @section Initialization:
 #' Currently support for directory roles is limited. Objects of this class should not be initialized directly.
@@ -37,10 +38,10 @@ public=list(
         super$initialize(token, tenant, properties)
     },
 
-    list_members=function()
+    list_members=function(n=Inf)
     {
-        res <- private$get_paged_list(self$do_operation("members"))
-        private$init_list_objects(res)
+        pager <- self$get_list_pager(self$do_operation("members"))
+        extract_list_values(pager, n)
     },
 
     print=function(...)