+++ R/ms_graph.R
@@ -19,6 +19,7 @@
 #' - `delete_group(group_id, confirm=TRUE)`: Deletes a group.
 #' - `call_graph_endpoint(op="", ...)`: Calls the Microsoft Graph API using this object's token and tenant as authentication arguments. See [call_graph_endpoint].
 #' - `call_batch_endpoint(requests=list(), ...)`: Calls the batch endpoint with a list of individual requests. See [call_batch_endpoint].
+#' - `get_aad_object(id)`: Retrieves an arbitrary Azure Active Directory object by ID.
 #'
 #' @section Authentication:
 #' The recommended way to authenticate with Microsoft Graph is via the [create_graph_login] function, which creates a new instance of this class.
@@ -53,7 +54,7 @@
 #' \dontrun{
 #'
 #' # start a new Graph session
-#' gr <- ms_graph$new(tenant="myaadtenant.onmicrosoft.com", app="app_id", password="password")
+#' gr <- ms_graph$new(tenant="myaadtenant.onmicrosoft.com")
 #'
 #' # authenticate with credentials in a file
 #' gr <- ms_graph$new(config_file="creds.json")
@@ -76,13 +77,17 @@
 #' cert <- readLines("mycert.cer")
 #' gr$create_app("mycertapp", password=FALSE, certificate=cert)
 #'
-#' # retrieving your own user details
+#' # retrieving your own user details (assuming interactive authentication)
 #' gr$get_user()
 #'
 #' # retrieving another user's details
 #' gr$get_user("username@myaadtenant.onmicrosoft.com")
 #' gr$get_user(email="firstname.lastname@mycompany.com")
 #' gr$get_user(name="Hong Ooi")
+#'
+#' # get an AAD object (a group)
+#' id <- gr$get_user()$list_group_memberships()[1]
+#' gr$get_aad_object(id)
 #'
 #' }
 #' @format An R6 object of class `ms_graph`.
@@ -297,6 +302,13 @@ public=list(
     call_batch_endpoint=function(requests=list(), ...)
     {
         call_batch_endpoint(self$token, requests, ...)
+    },
+
+    get_aad_object=function(id)
+    {
+        res <- self$call_graph_endpoint(file.path("directoryObjects", id))
+        gen <- find_class_generator(res, type_filter=NULL)
+        gen$new(self$token, self$tenant, res)
     },
 
     print=function(...)