+++ R/ms_object.R
@@ -1,4 +1,4 @@
-#' Azure Active Directory object
+#' Microsoft Graph object
 #'
 #' Base class representing a object in Microsoft Graph. All other Graph object classes ultimately inherit from this class.
 #'
@@ -14,25 +14,38 @@
 #' - `update(...)`: Update the object information in Azure Active Directory.
 #' - `do_operation(...)`: Carry out an arbitrary operation on the object.
 #' - `sync_fields()`: Synchronise the R object with the data in Azure Active Directory.
-#'
-#' The following methods are private, and are intended for package authors extending AzureGraph to define their own objects.
-#' - `get_paged_list(lst, next_link_name, value_name, simplify, n)`: Used to process API calls that return lists of objects. Microsoft Graph returns lists in pages, with each page containing a subset of objects and a link to the next page. This method reconstructs the list, given the first page. Its arguments are:
-#'    - `lst`: Object containing the initial page of results, generally the result of a call to `do_operation`. Should be a list with names corresponding to the arguments `next_link_name` and `value_name`.
-#'    - `next_link_name`: The name of the component of `lst` containing the link to the next page. Defaults to `@odata.nextLink`.
-#'    - `value_name`: The name of the component of `lst` containing the first page of results. Defaults to `value`.
-#'    - `simplify`: Whether to turn the list of objects into a data frame. This is useful if the list is intended to be a rectangular data structure, eg a SharePoint list or OneDrive file listing. Note that the vctrs package must be installed to return a data frame, if the objects in the list can have varying structures (which will often be the case).
-#'    - `n`: Optionally, limit the list to this many objects.
-#' - `init_list_objects(lst, type_filter, default_generator, ...)`: `get_paged_list` returns a raw list, the result of parsing the JSON response from the Graph host. This method converts the list into actual R6 objects. Its arguments are:
-#'   - `lst`: The input list.
-#'   - `type_filter`: The possible types of objects that the list contains. The default is NULL.
-#'   - `default_generator`: An R6 class generator object to use, if `init_list_objects` is unable to detect the type of an object. It's recommended to change this from the default value of `ms_object`, if you know that all objects in the list will have the same class.
-#'   - `...`: Further arguments to pass to the generator's `initialize` method.
+#' - `get_list_pager(...)`: Returns a pager object, which is an _iterator_ for a set of paged query results. See 'Paged results' below.
 #'
 #' @section Initialization:
 #' Objects of this class should not be created directly. Instead, create an object of the appropriate subclass.
 #'
+#' @section Paged results:
+#' Microsoft Graph returns lists in pages, with each page containing a subset of objects and a link to the next page. AzureGraph provides an iterator-based API that lets you access each page individually, or collect them all into a single object.
+#'
+#' To create a new pager object, call the `get_list_pager()` method with the following arguments:
+#' - `lst`: A list containing the first page of results, generally from a call to the `do_operation()` method.
+#' - `next_link_name,value_name`: The names of the components of `first_page` containing the link to the next page, and the set of values for the page respectively. The default values are `@odata.nextLink` and `value`.
+#' - `generate_objects`: Whether the iterator should return a list containing the parsed JSON for the page values, or convert it into a list of R6 objects.
+#' - `type_filter`: Any extra arguments required to initialise the returned objects. Only used if `generate_objects` is TRUE.
+#' - `...`: Any extra arguments required to initialise the returned objects. Only used if `generate_objects` is TRUE.
+#'
+#' This returns an object of class [ms_graph_pager], which is an _iterator_ for the set of paged results. Each call to the object's `value` active binding yields the next page. When all pages have been returned, `value` contains NULL.
+#'
+#' The format of the returned values can take one of 3 forms, based on the initial format of the first page and the `generate_objects` argument.
+#'
+#' If the first page of results is a data frame (each item has been converted into a row), then the pager will return results as data frames. In this case, the `output` field is automatically set to "data.frame" and the `generate_objects` initialization argument is ignored. Usually this will be the case when the results are meant to represent external data, eg items in a SharePoint list.
+#'
+#' If the first page of results is a list, the `generate_objects` argument sets whether to convert the items in each page into R6 objects defined by the AzureGraph class framework. If `generate_objects` is TRUE, the `output` field is set to "object", and if `generate_objects` is FALSE, the `output` field is set to "list".
+#'
+#' You can also call the `extract_list_values()` function to get all or some of the values from a pager, without having to manually combine the pages together.
+#'
+#' @section Deprecated methods:
+#' The following methods are private and **deprecated**, and form the older AzureGraph API for accessing paged results. They will eventually be removed.
+#' - `get_paged_list(lst, next_link_name, value_name, simplify, n)`: This method reconstructs the list, given the first page.
+#' - `init_list_objects(lst, type_filter, default_generator, ...)`: `get_paged_list` returns a raw list, the result of parsing the JSON response from the Graph host. This method converts the list into actual R6 objects.
+#'
 #' @seealso
-#' [ms_graph], [az_object]
+#' [ms_graph], [az_object], [ms_graph_pager], [extract_list_values]
 #'
 #' [Microsoft Graph overview](https://docs.microsoft.com/en-us/graph/overview),
 #' [REST API reference](https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-1.0)
@@ -95,6 +108,12 @@ public=list(
     {
         op <- construct_path(private$api_type, self$properties$id, op)
         call_graph_endpoint(self$token, op, ...)
+    },
+
+    get_list_pager=function(lst, next_link_name="@odata.nextLink", value_name="value", generate_objects=TRUE,
+                            type_filter=NULL, ...)
+    {
+        ms_graph_pager$new(self$token, lst, next_link_name, value_name, generate_objects, type_filter, ...)
     },
 
     print=function(...)