+++ NEWS.md
@@ -1,7 +1,16 @@
 # AzureGraph 1.2.2.9000
 
+- New API for working with paged result sets:
+  - New `ms_graph_pager` R6 class, which is an _iterator_ for the pages in the result.
+  - The `ms_object` base class now has a `get_list_pager()` method which returns an object of class `ms_graph_pager`.
+  - New `extract_list_values()` function to get all or part of the results from a paged result set.
+- The current (private) `ms_object$get_paged_list()` and `ms_object$init_list_objects()` methods are retained for backward compatibility, but are otherwise deprecated.
 - The `ms_graph$get_user()` method can now get a user by email or display name.
 - Fix a bug in retrieving a paged list of values as a data frame, when `n` (the maximum number of rows) is supplied.
+- New `ms_graph$get_aad_object()` method to retrieve an Azure Active Directory object by ID. Mostly intended for use with the `list_object_memberships()` and `list_group_memberships()` methods, which return only IDs and not full object information.
+- All `list_*` methods now have an `n` argument to set a cap on the number of results; the default value is `n=Inf`. If this is set to NULL, the `ms_graph_pager` iterator object is returned instead to allow manual iteration over the results.
+- Export the `find_class_generator()` function.
+- New "Batching and paging" vignette describing these APIs.
 
 # AzureGraph 1.2.2